Revert "Squash merge from master"
This reverts commit 51242a1d57504039a4407b52db54680c3589fa3a.

+++ com.unity.render-pipelines.universal/Editor/ShaderGraph/Includes/SpriteLitPass.hlsl
@@ -1,6 +1,6 @@
 #if ETC1_EXTERNAL_ALPHA
     TEXTURE2D(_AlphaTex); SAMPLER(sampler_AlphaTex);
-    half _EnableAlphaTexture;
+    float _EnableAlphaTexture;
 #endif
 
 #if USE_SHAPE_LIGHT_TYPE_0
@@ -39,7 +39,7 @@ half4 frag(PackedVaryings packedInput) : SV_TARGET
     SurfaceDescription surfaceDescription = SurfaceDescriptionFunction(surfaceDescriptionInputs);
 
 #if ETC1_EXTERNAL_ALPHA
-    half4 alpha = SAMPLE_TEXTURE2D(_AlphaTex, sampler_AlphaTex, unpacked.texCoord0.xy);
+    float4 alpha = SAMPLE_TEXTURE2D(_AlphaTex, sampler_AlphaTex, unpacked.texCoord0.xy);
     surfaceDescription.Color.a = lerp (surfaceDescription.Color.a, alpha.r, _EnableAlphaTexture);
 #endif
