+++ com.unity.render-pipelines.universal/Shaders/2D/Include/LightingUtility.hlsl
@@ -1,8 +1,8 @@
 #if USE_NORMAL_MAP
     #if LIGHT_QUALITY_FAST
         #define NORMALS_LIGHTING_COORDS(TEXCOORDA, TEXCOORDB) \
-            half4	lightDirection	: TEXCOORDA;\
-            half2	screenUV   : TEXCOORDB;
+            float4	lightDirection	: TEXCOORDA;\
+            float2	screenUV   : TEXCOORDB;
 
         #define TRANSFER_NORMALS_LIGHTING(output, worldSpacePos)\
             float4 clipVertex = output.positionCS / output.positionCS.w;\
@@ -14,12 +14,12 @@
             
         #define APPLY_NORMALS_LIGHTING(input, lightColor)\
             half4 normal = SAMPLE_TEXTURE2D(_NormalMap, sampler_NormalMap, input.screenUV);\
-            half3 normalUnpacked = UnpackNormal(normal);\
+            float3 normalUnpacked = UnpackNormal(normal);\
             lightColor = lightColor * saturate(dot(input.lightDirection.xyz, normalUnpacked));
     #else
         #define NORMALS_LIGHTING_COORDS(TEXCOORDA, TEXCOORDB) \
-            half4	positionWS : TEXCOORDA;\
-            half2	screenUV   : TEXCOORDB;
+            float4	positionWS : TEXCOORDA;\
+            float2	screenUV   : TEXCOORDB;
 
         #define TRANSFER_NORMALS_LIGHTING(output, worldSpacePos) \
             float4 clipVertex = output.positionCS / output.positionCS.w;\
@@ -28,8 +28,8 @@
 
         #define APPLY_NORMALS_LIGHTING(input, lightColor)\
             half4 normal = SAMPLE_TEXTURE2D(_NormalMap, sampler_NormalMap, input.screenUV);\
-            half3 normalUnpacked = UnpackNormal(normal);\
-            half3 dirToLight;\
+            float3 normalUnpacked = UnpackNormal(normal);\
+            float3 dirToLight;\
             dirToLight.xy = _LightPosition.xy - input.positionWS.xy;\
             dirToLight.z =  _LightZDistance;\
             dirToLight = normalize(dirToLight);\
@@ -39,7 +39,7 @@
     #define NORMALS_LIGHTING_VARIABLES \
             TEXTURE2D(_NormalMap); \
             SAMPLER(sampler_NormalMap); \
-            half4	    _LightPosition;\
+            float4	_LightPosition;\
             half	    _LightZDistance;
 #else
     #define NORMALS_LIGHTING_COORDS(TEXCOORDA, TEXCOORDB)
@@ -73,7 +73,7 @@
 #define SHAPE_LIGHT(index)\
     TEXTURE2D(_ShapeLightTexture##index);\
     SAMPLER(sampler_ShapeLightTexture##index);\
-    half2 _ShapeLightBlendFactors##index;\
-    half4 _ShapeLightMaskFilter##index;\
-    half4 _ShapeLightInvertedFilter##index;
+    float2 _ShapeLightBlendFactors##index;\
+    float4 _ShapeLightMaskFilter##index;\
+    float4 _ShapeLightInvertedFilter##index;
